# inventory-be/docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Automation and Tooling API",
    "description": "Comprehensive API documentation for the structural inventory platform. Includes Auth, RBAC, Users, Locations, Stocks, Requests, Audit, Notification, and Dashboard modules.",
    "version": "1.1.0"
  },
  "servers": [
    {
      "url": "http://localhost:3000/",
      "description": "Local Development Server"
    }
  ],
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string",
            "example": "Resource not found."
          }
        }
      },
      "PaginationMeta": {
        "type": "object",
        "properties": {
          "total": { "type": "integer", "example": 100 },
          "page": { "type": "integer", "example": 1 },
          "limit": { "type": "integer", "example": 10 },
          "totalPages": { "type": "integer", "example": 10 }
        }
      },
...
      "Role": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "name": { "type": "string", "example": "admin" },
          "description": { "type": "string", "example": "Administrator role" },
          "createdAt": { "type": "string", "format": "date-time" },
          "updatedAt": { "type": "string", "format": "date-time" }
        }
      },
      "RoleWithPermissions": {
        "type": "object",
        "allOf": [{ "$ref": "#/components/schemas/Role" }],
        "properties": {
          "permissions": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/Permission" }
          }
        }
      },
      "Permission": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "name": { "type": "string", "example": "stocks:write" },
          "description": { "type": "string", "example": "Can create and mutate stocks." },
          "createdAt": { "type": "string", "format": "date-time" },
          "updatedAt": { "type": "string", "format": "date-time" }
        }
      },
...
    },
    "parameters": {
      "PageParam": {
        "name": "page", "in": "query",
        "schema": { "type": "integer", "default": 1 }
      },
      "LimitParam": {
        "name": "limit", "in": "query",
        "schema": { "type": "integer", "default": 10 }
      },
      "SearchParam": {
        "name": "search", "in": "query",
        "schema": { "type": "string" }
      },
      "IdPathParam": {
        "name": "id", "in": "path", "required": true,
        "schema": { "type": "string", "format": "uuid" }
      }
    }
  },
  "security": [{ "BearerAuth": [] }],
  "paths": {
...
    "/rbac/roles": {
      "get": {
        "summary": "Get all Roles — paginated (Requires any rbac permission)",
        "tags": ["RBAC"],
        "parameters": [
          { "$ref": "#/components/parameters/PageParam" },
          { "$ref": "#/components/parameters/LimitParam" },
          { "$ref": "#/components/parameters/SearchParam" }
        ],
        "responses": {
          "200": {
            "description": "Paginated roles list",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": { "type": "array", "items": { "$ref": "#/components/schemas/Role" } },
                    "meta": { "$ref": "#/components/schemas/PaginationMeta" }
                  }
                }
              }
            }
          }
        },
        "operationId": "getRbacRoles"
      },
      "post": {
        "summary": "Create Role (Requires any rbac permission)",
        "tags": ["RBAC"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["name"],
                "properties": {
                  "name": { "type": "string" },
                  "description": { "type": "string" }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Role created",
            "content": { "application/json": { "schema": { "type": "object", "properties": { "data": { "$ref": "#/components/schemas/Role" } } } } }
          },
          "409": { "description": "Role name already exists", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } }
        },
        "operationId": "postRbacRoles"
      }
    },
    "/rbac/roles/{id}": {
      "get": {
        "summary": "Get Role by ID with its assigned permissions (Requires any rbac permission)",
        "tags": ["RBAC"],
        "parameters": [{ "$ref": "#/components/parameters/IdPathParam" }],
        "responses": {
          "200": {
            "description": "Role with permissions",
            "content": { "application/json": { "schema": { "type": "object", "properties": { "data": { "$ref": "#/components/schemas/RoleWithPermissions" } } } } }
          },
          "404": { "description": "Role not found", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } }
        },
        "operationId": "getRbacRolesId"
      },
      "patch": {
        "summary": "Partially update a Role (Requires any rbac permission)",
        "tags": ["RBAC"],
        "parameters": [{ "$ref": "#/components/parameters/IdPathParam" }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": { "type": "string" },
                  "description": { "type": "string" }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Role updated", "content": { "application/json": { "schema": { "type": "object", "properties": { "data": { "$ref": "#/components/schemas/Role" } } } } } },
          "404": { "description": "Role not found", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } },
          "409": { "description": "Role name already exists", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } }
        },
        "operationId": "patchRbacRolesId"
      },
      "delete": {
        "summary": "Delete a Role (Requires any rbac permission)",
        "tags": ["RBAC"],
        "parameters": [{ "$ref": "#/components/parameters/IdPathParam" }],
        "responses": {
          "200": { "description": "Role deleted", "content": { "application/json": { "schema": { "type": "object", "properties": { "data": { "$ref": "#/components/schemas/Role" } } } } } },
          "404": { "description": "Role not found", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } }
        },
        "operationId": "deleteRbacRolesId"
      }
    },
    "/rbac/permissions": {
      "get": {
        "summary": "Get all Permissions — paginated (Requires any rbac permission)",
        "tags": ["RBAC"],
        "parameters": [
          { "$ref": "#/components/parameters/PageParam" },
          { "$ref": "#/components/parameters/LimitParam" },
          { "$ref": "#/components/parameters/SearchParam" }
        ],
        "responses": {
          "200": {
            "description": "Paginated permissions list",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": { "type": "array", "items": { "$ref": "#/components/schemas/Permission" } },
                    "meta": { "$ref": "#/components/schemas/PaginationMeta" }
                  }
                }
              }
            }
          }
        },
        "operationId": "getRbacPermissions"
      },
      "post": {
        "summary": "Create Permission (Requires any rbac permission)",
        "tags": ["RBAC"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["name"],
                "properties": {
                  "name": { "type": "string", "example": "stocks:write" },
                  "description": { "type": "string" }
                }
              }
            }
          }
        },
        "responses": {
          "201": { "description": "Permission created", "content": { "application/json": { "schema": { "type": "object", "properties": { "data": { "$ref": "#/components/schemas/Permission" } } } } } },
          "409": { "description": "Permission name already exists", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } }
        },
        "operationId": "postRbacPermissions"
      }
    },
    "/rbac/permissions/{id}": {
      "patch": {
        "summary": "Partially update a Permission (Requires any rbac permission)",
        "tags": ["RBAC"],
        "parameters": [{ "$ref": "#/components/parameters/IdPathParam" }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": { "type": "string" },
                  "description": { "type": "string" }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Permission updated", "content": { "application/json": { "schema": { "type": "object", "properties": { "data": { "$ref": "#/components/schemas/Permission" } } } } } },
          "404": { "description": "Permission not found", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } },
          "409": { "description": "Permission name already exists", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } }
        },
        "operationId": "patchRbacPermissionsId"
      },
      "delete": {
        "summary": "Delete a Permission (Requires any rbac permission)",
        "tags": ["RBAC"],
        "parameters": [{ "$ref": "#/components/parameters/IdPathParam" }],
        "responses": {
          "200": { "description": "Permission deleted", "content": { "application/json": { "schema": { "type": "object", "properties": { "data": { "$ref": "#/components/schemas/Permission" } } } } } },
          "404": { "description": "Permission not found", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } }
        },
        "operationId": "deleteRbacPermissionsId"
      }
    },
    "/rbac/roles/assign-permission": {
      "post": {
        "summary": "Assign a Permission to a Role (Requires any rbac permission)",
        "tags": ["RBAC"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["roleId", "permissionId"],
                "properties": {
                  "roleId": { "type": "string", "format": "uuid" },
                  "permissionId": { "type": "string", "format": "uuid" }
                }
              }
            }
          }
        },
        "responses": {
          "201": { "description": "Permission assigned to role" },
          "400": { "description": "Bad Request", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } }
        },
        "operationId": "postRbacRolesAssignPermission"
      }
    },
    "/rbac/roles/remove-permission": {
      "post": {
        "summary": "Remove a Permission from a Role (Requires any rbac permission)",
        "tags": ["RBAC"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["roleId", "permissionId"],
                "properties": {
                  "roleId": { "type": "string", "format": "uuid" },
                  "permissionId": { "type": "string", "format": "uuid" }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Permission removed from role" },
          "400": { "description": "Bad Request", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } }
        },
        "operationId": "postRbacRolesRemovePermission"
      }
    },
    "/rbac/users/assign-role": {
      "post": {
        "summary": "Assign a Role to a User (Requires any rbac permission)",
        "tags": ["RBAC"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["userId", "roleId"],
                "properties": {
                  "userId": { "type": "string", "format": "uuid" },
                  "roleId": { "type": "string", "format": "uuid" }
                }
              }
            }
          }
        },
        "responses": {
          "201": { "description": "Role assigned to user" },
          "400": { "description": "Bad Request", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } }
        },
        "operationId": "postRbacUsersAssignRole"
      }
    },
...
  }
}